// types/ve3/connector-alliance/instantiate.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * Represents the type of an fungible asset.
 *
 * Each **asset info** instance can be one of three variants:
 *
 * - Native SDK coins. To create an **asset info** instance of this type, provide the denomination. - CW20 tokens. To create an **asset info** instance of this type, provide the contract address.
 */
export type AssetInfoBaseFor_String =
  | {
      native: string;
    }
  | {
      cw20: string;
    };

export interface InstantiateMsg {
  alliance_token_denom: string;
  gauge: string;
  global_config_addr: string;
  lst_asset_info: AssetInfoBaseFor_String;
  lst_hub_address: string;
  reward_denom: string;
  zasset_denom: string;
}